Move Celery tasks from h.indexer to h.tasks.indexer
This one is a little more fiddly than the prior commits, because
h.indexer contained both Celery tasks and integration between the h
application and memex's annotation events.

This commit:

- moves the Celery tasks into h.tasks.indexer
- removes the unused "reindex" task
